test(tools): split handlers_test.go under 500 LOC; add registry_test.go (deslop #39)
Pure move of test functions, no body changes:
- registry_test.go: TestGetHandler, TestUnknownToolHandler (registry.go
  dispatch/lookup machinery, gives registry.go its matching test file)
- handlers_search_test.go: search + autocomplete handler tests
- handlers_lookup_test.go: relationship, issuer, reporting-exception tests
- handlers_test.go: 1016 -> 435 LOC (helpers + lookup/validate/batch + helper tests)

All four test files now under 500 LOC. Shared fixtures stay in
handlers_test.go (same package, referenced without import).
