+++ tools/handlers_test.go
@@ -312,587 +312,6 @@ func TestHandleBatchLEILookup(t *testing.T) {
 	})
 }
 
-// TestHandleSearchEntity tests the search_entity handler.
-func TestHandleSearchEntity(t *testing.T) {
-	t.Run("successful search", func(t *testing.T) {
-		ts := newTestServer()
-		defer ts.Close()
-
-		ts.mux.HandleFunc("/lei-records", func(w http.ResponseWriter, r *http.Request) {
-			resp := mockSearchResponse(
-				gleif.LEIRecord{
-					LEI:    "HWUPKR0MPOU8FGXBT394",
-					Entity: gleif.Entity{LegalName: gleif.LegalName{Name: "Apple Inc."}},
-				},
-			)
-			w.Header().Set("Content-Type", "application/vnd.api+json")
-			_ = json.NewEncoder(w).Encode(resp)
-		})
-
-		registry := newTestRegistry(ts.URL())
-		req := makeRequest(map[string]any{"query": "Apple"})
-
-		result, err := registry.handleSearchEntity(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleSearchEntity returned error: %v", err)
-		}
-		if result.IsError {
-			t.Fatalf("handleSearchEntity returned error result: %s", getResultText(result))
-		}
-
-		text := getResultText(result)
-		if !strings.Contains(text, "Apple Inc.") {
-			t.Errorf("Expected result to contain 'Apple Inc.', got: %s", text)
-		}
-	})
-
-	t.Run("with pagination parameters", func(t *testing.T) {
-		ts := newTestServer()
-		defer ts.Close()
-
-		ts.mux.HandleFunc("/lei-records", func(w http.ResponseWriter, r *http.Request) {
-			resp := mockSearchResponse()
-			w.Header().Set("Content-Type", "application/vnd.api+json")
-			_ = json.NewEncoder(w).Encode(resp)
-		})
-
-		registry := newTestRegistry(ts.URL())
-		req := makeRequest(map[string]any{
-			"query": "Test",
-			"limit": float64(10),
-			"page":  float64(2),
-			"fuzzy": false,
-		})
-
-		result, err := registry.handleSearchEntity(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleSearchEntity returned error: %v", err)
-		}
-		if result.IsError {
-			t.Fatalf("handleSearchEntity returned error result: %s", getResultText(result))
-		}
-	})
-
-	t.Run("missing query parameter", func(t *testing.T) {
-		registry := newTestRegistry("http://unused")
-		req := makeRequest(map[string]any{})
-
-		result, err := registry.handleSearchEntity(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleSearchEntity returned error: %v", err)
-		}
-		if !result.IsError {
-			t.Error("Expected error result for missing query parameter")
-		}
-	})
-}
-
-// TestHandleSearchByBIC tests the search_by_bic handler.
-func TestHandleSearchByBIC(t *testing.T) {
-	t.Run("successful search", func(t *testing.T) {
-		ts := newTestServer()
-		defer ts.Close()
-
-		ts.mux.HandleFunc("/lei-records", func(w http.ResponseWriter, r *http.Request) {
-			resp := mockSearchResponse(
-				gleif.LEIRecord{
-					LEI:    "7LTWFZYICNSX8D621K86",
-					Entity: gleif.Entity{LegalName: gleif.LegalName{Name: "Deutsche Bank AG"}},
-				},
-			)
-			w.Header().Set("Content-Type", "application/vnd.api+json")
-			_ = json.NewEncoder(w).Encode(resp)
-		})
-
-		registry := newTestRegistry(ts.URL())
-		req := makeRequest(map[string]any{"bic": "DEUTDEFF"})
-
-		result, err := registry.handleSearchByBIC(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleSearchByBIC returned error: %v", err)
-		}
-		if result.IsError {
-			t.Fatalf("handleSearchByBIC returned error result: %s", getResultText(result))
-		}
-
-		text := getResultText(result)
-		if !strings.Contains(text, `"found": true`) {
-			t.Errorf("Expected found=true in result, got: %s", text)
-		}
-	})
-
-	t.Run("no results", func(t *testing.T) {
-		ts := newTestServer()
-		defer ts.Close()
-
-		ts.mux.HandleFunc("/lei-records", func(w http.ResponseWriter, r *http.Request) {
-			resp := mockSearchResponse()
-			w.Header().Set("Content-Type", "application/vnd.api+json")
-			_ = json.NewEncoder(w).Encode(resp)
-		})
-
-		registry := newTestRegistry(ts.URL())
-		req := makeRequest(map[string]any{"bic": "TESTTEST"})
-
-		result, err := registry.handleSearchByBIC(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleSearchByBIC returned error: %v", err)
-		}
-		if result.IsError {
-			t.Fatalf("handleSearchByBIC returned error result: %s", getResultText(result))
-		}
-
-		text := getResultText(result)
-		if !strings.Contains(text, `"found": false`) {
-			t.Errorf("Expected found=false in result, got: %s", text)
-		}
-	})
-
-	t.Run("invalid bic length", func(t *testing.T) {
-		registry := newTestRegistry("http://unused")
-		req := makeRequest(map[string]any{"bic": "TOOLONG1234567890"})
-
-		result, err := registry.handleSearchByBIC(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleSearchByBIC returned error: %v", err)
-		}
-		if !result.IsError {
-			t.Error("Expected error result for invalid bic length")
-		}
-	})
-
-	t.Run("missing bic parameter", func(t *testing.T) {
-		registry := newTestRegistry("http://unused")
-		req := makeRequest(map[string]any{})
-
-		result, err := registry.handleSearchByBIC(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleSearchByBIC returned error: %v", err)
-		}
-		if !result.IsError {
-			t.Error("Expected error result for missing bic parameter")
-		}
-	})
-}
-
-// TestHandleSearchByISIN tests the search_by_isin handler.
-func TestHandleSearchByISIN(t *testing.T) {
-	t.Run("successful search", func(t *testing.T) {
-		ts := newTestServer()
-		defer ts.Close()
-
-		ts.mux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-			resp := mockSearchResponse(
-				gleif.LEIRecord{
-					LEI:    "HWUPKR0MPOU8FGXBT394",
-					Entity: gleif.Entity{LegalName: gleif.LegalName{Name: "Apple Inc."}},
-				},
-			)
-			w.Header().Set("Content-Type", "application/vnd.api+json")
-			_ = json.NewEncoder(w).Encode(resp)
-		})
-
-		registry := newTestRegistry(ts.URL())
-		req := makeRequest(map[string]any{"isin": "US0378331005"})
-
-		result, err := registry.handleSearchByISIN(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleSearchByISIN returned error: %v", err)
-		}
-		if result.IsError {
-			t.Fatalf("handleSearchByISIN returned error result: %s", getResultText(result))
-		}
-	})
-
-	t.Run("invalid isin length", func(t *testing.T) {
-		registry := newTestRegistry("http://unused")
-		req := makeRequest(map[string]any{"isin": "TOOLONG1234567890"})
-
-		result, err := registry.handleSearchByISIN(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleSearchByISIN returned error: %v", err)
-		}
-		if !result.IsError {
-			t.Error("Expected error result for invalid isin length")
-		}
-	})
-
-	t.Run("missing isin parameter", func(t *testing.T) {
-		registry := newTestRegistry("http://unused")
-		req := makeRequest(map[string]any{})
-
-		result, err := registry.handleSearchByISIN(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleSearchByISIN returned error: %v", err)
-		}
-		if !result.IsError {
-			t.Error("Expected error result for missing isin parameter")
-		}
-	})
-}
-
-// TestHandleSearchByCountry tests the search_by_country handler.
-func TestHandleSearchByCountry(t *testing.T) {
-	t.Run("successful search", func(t *testing.T) {
-		ts := newTestServer()
-		defer ts.Close()
-
-		ts.mux.HandleFunc("/lei-records", func(w http.ResponseWriter, r *http.Request) {
-			resp := mockSearchResponse(
-				gleif.LEIRecord{
-					LEI:    "HWUPKR0MPOU8FGXBT394",
-					Entity: gleif.Entity{LegalName: gleif.LegalName{Name: "Apple Inc."}},
-				},
-			)
-			w.Header().Set("Content-Type", "application/vnd.api+json")
-			_ = json.NewEncoder(w).Encode(resp)
-		})
-
-		registry := newTestRegistry(ts.URL())
-		req := makeRequest(map[string]any{"country": "US"})
-
-		result, err := registry.handleSearchByCountry(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleSearchByCountry returned error: %v", err)
-		}
-		if result.IsError {
-			t.Fatalf("handleSearchByCountry returned error result: %s", getResultText(result))
-		}
-
-		text := getResultText(result)
-		if !strings.Contains(text, `"country": "US"`) {
-			t.Errorf("Expected country=US in result, got: %s", text)
-		}
-	})
-
-	t.Run("invalid country code", func(t *testing.T) {
-		registry := newTestRegistry("http://unused")
-		req := makeRequest(map[string]any{"country": "USA"})
-
-		result, err := registry.handleSearchByCountry(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleSearchByCountry returned error: %v", err)
-		}
-		if !result.IsError {
-			t.Error("Expected error result for invalid country code")
-		}
-	})
-
-	t.Run("missing country parameter", func(t *testing.T) {
-		registry := newTestRegistry("http://unused")
-		req := makeRequest(map[string]any{})
-
-		result, err := registry.handleSearchByCountry(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleSearchByCountry returned error: %v", err)
-		}
-		if !result.IsError {
-			t.Error("Expected error result for missing country parameter")
-		}
-	})
-}
-
-// TestHandleGetRelationships tests the get_relationships handler.
-func TestHandleGetRelationships(t *testing.T) {
-	t.Run("successful lookup", func(t *testing.T) {
-		ts := newTestServer()
-		defer ts.Close()
-
-		ts.mux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-			resp := struct {
-				Data []any `json:"data"`
-			}{Data: []any{}}
-			w.Header().Set("Content-Type", "application/vnd.api+json")
-			_ = json.NewEncoder(w).Encode(resp)
-		})
-
-		registry := newTestRegistry(ts.URL())
-		req := makeRequest(map[string]any{"lei": "HWUPKR0MPOU8FGXBT394"})
-
-		result, err := registry.handleGetRelationships(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleGetRelationships returned error: %v", err)
-		}
-		if result.IsError {
-			t.Fatalf("handleGetRelationships returned error result: %s", getResultText(result))
-		}
-	})
-
-	t.Run("invalid lei format", func(t *testing.T) {
-		registry := newTestRegistry("http://unused")
-		req := makeRequest(map[string]any{"lei": "INVALID"})
-
-		result, err := registry.handleGetRelationships(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleGetRelationships returned error: %v", err)
-		}
-		if !result.IsError {
-			t.Error("Expected error result for invalid lei format")
-		}
-	})
-
-	t.Run("missing lei parameter", func(t *testing.T) {
-		registry := newTestRegistry("http://unused")
-		req := makeRequest(map[string]any{})
-
-		result, err := registry.handleGetRelationships(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleGetRelationships returned error: %v", err)
-		}
-		if !result.IsError {
-			t.Error("Expected error result for missing lei parameter")
-		}
-	})
-}
-
-// TestHandleAutocomplete tests the autocomplete handler.
-func TestHandleAutocomplete(t *testing.T) {
-	t.Run("successful autocomplete via fallback", func(t *testing.T) {
-		ts := newTestServer()
-		defer ts.Close()
-
-		ts.mux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-			resp := mockSearchResponse(
-				gleif.LEIRecord{
-					LEI:    "HWUPKR0MPOU8FGXBT394",
-					Entity: gleif.Entity{LegalName: gleif.LegalName{Name: "Apple Inc."}},
-				},
-			)
-			w.Header().Set("Content-Type", "application/vnd.api+json")
-			_ = json.NewEncoder(w).Encode(resp)
-		})
-
-		registry := newTestRegistry(ts.URL())
-		req := makeRequest(map[string]any{"prefix": "Apple"})
-
-		result, err := registry.handleAutocomplete(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleAutocomplete returned error: %v", err)
-		}
-		if result.IsError {
-			t.Fatalf("handleAutocomplete returned error result: %s", getResultText(result))
-		}
-
-		text := getResultText(result)
-		if !strings.Contains(text, "suggestions") {
-			t.Errorf("Expected result to contain 'suggestions', got: %s", text)
-		}
-	})
-
-	t.Run("missing prefix parameter", func(t *testing.T) {
-		registry := newTestRegistry("http://unused")
-		req := makeRequest(map[string]any{})
-
-		result, err := registry.handleAutocomplete(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleAutocomplete returned error: %v", err)
-		}
-		if !result.IsError {
-			t.Error("Expected error result for missing prefix parameter")
-		}
-	})
-}
-
-// TestHandleGetLEIIssuer tests the get_lei_issuer handler.
-func TestHandleGetLEIIssuer(t *testing.T) {
-	t.Run("successful lookup", func(t *testing.T) {
-		ts := newTestServer()
-		defer ts.Close()
-
-		ts.mux.HandleFunc("/lei-issuers/", func(w http.ResponseWriter, r *http.Request) {
-			resp := gleif.SingleResponse[gleif.LEIIssuer]{
-				Data: gleif.DataItem[gleif.LEIIssuer]{
-					ID:   "EVK05KS7XY1DEII3R011",
-					Type: "lei-issuers",
-					Attributes: gleif.LEIIssuer{
-						Name:    "WM Datenservice",
-						Country: "DE",
-						Status:  "ACTIVE",
-					},
-				},
-			}
-			w.Header().Set("Content-Type", "application/vnd.api+json")
-			_ = json.NewEncoder(w).Encode(resp)
-		})
-
-		registry := newTestRegistry(ts.URL())
-		req := makeRequest(map[string]any{"issuer_id": "EVK05KS7XY1DEII3R011"})
-
-		result, err := registry.handleGetLEIIssuer(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleGetLEIIssuer returned error: %v", err)
-		}
-		if result.IsError {
-			t.Fatalf("handleGetLEIIssuer returned error result: %s", getResultText(result))
-		}
-
-		text := getResultText(result)
-		if !strings.Contains(text, "WM Datenservice") {
-			t.Errorf("Expected result to contain 'WM Datenservice', got: %s", text)
-		}
-	})
-
-	t.Run("missing issuer_id parameter", func(t *testing.T) {
-		registry := newTestRegistry("http://unused")
-		req := makeRequest(map[string]any{})
-
-		result, err := registry.handleGetLEIIssuer(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleGetLEIIssuer returned error: %v", err)
-		}
-		if !result.IsError {
-			t.Error("Expected error result for missing issuer_id parameter")
-		}
-	})
-}
-
-// TestHandleListLEIIssuers tests the list_lei_issuers handler.
-func TestHandleListLEIIssuers(t *testing.T) {
-	t.Run("successful list", func(t *testing.T) {
-		ts := newTestServer()
-		defer ts.Close()
-
-		ts.mux.HandleFunc("/lei-issuers", func(w http.ResponseWriter, r *http.Request) {
-			resp := gleif.APIResponse[gleif.LEIIssuer]{
-				Data: []gleif.DataItem[gleif.LEIIssuer]{
-					{
-						ID:   "EVK05KS7XY1DEII3R011",
-						Type: "lei-issuers",
-						Attributes: gleif.LEIIssuer{
-							Name:    "WM Datenservice",
-							Country: "DE",
-						},
-					},
-				},
-			}
-			w.Header().Set("Content-Type", "application/vnd.api+json")
-			_ = json.NewEncoder(w).Encode(resp)
-		})
-
-		registry := newTestRegistry(ts.URL())
-		req := makeRequest(map[string]any{})
-
-		result, err := registry.handleListLEIIssuers(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleListLEIIssuers returned error: %v", err)
-		}
-		if result.IsError {
-			t.Fatalf("handleListLEIIssuers returned error result: %s", getResultText(result))
-		}
-
-		text := getResultText(result)
-		if !strings.Contains(text, `"count": 1`) {
-			t.Errorf("Expected count=1 in result, got: %s", text)
-		}
-	})
-}
-
-// TestHandleGetReportingExceptions tests the get_reporting_exceptions handler.
-func TestHandleGetReportingExceptions(t *testing.T) {
-	t.Run("successful lookup", func(t *testing.T) {
-		ts := newTestServer()
-		defer ts.Close()
-
-		ts.mux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-			resp := gleif.APIResponse[gleif.ReportingException]{
-				Data: []gleif.DataItem[gleif.ReportingException]{},
-			}
-			w.Header().Set("Content-Type", "application/vnd.api+json")
-			_ = json.NewEncoder(w).Encode(resp)
-		})
-
-		registry := newTestRegistry(ts.URL())
-		req := makeRequest(map[string]any{"lei": "HWUPKR0MPOU8FGXBT394"})
-
-		result, err := registry.handleGetReportingExceptions(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleGetReportingExceptions returned error: %v", err)
-		}
-		if result.IsError {
-			t.Fatalf("handleGetReportingExceptions returned error result: %s", getResultText(result))
-		}
-	})
-
-	t.Run("invalid lei format", func(t *testing.T) {
-		registry := newTestRegistry("http://unused")
-		req := makeRequest(map[string]any{"lei": "INVALID"})
-
-		result, err := registry.handleGetReportingExceptions(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleGetReportingExceptions returned error: %v", err)
-		}
-		if !result.IsError {
-			t.Error("Expected error result for invalid lei format")
-		}
-	})
-
-	t.Run("missing lei parameter", func(t *testing.T) {
-		registry := newTestRegistry("http://unused")
-		req := makeRequest(map[string]any{})
-
-		result, err := registry.handleGetReportingExceptions(context.Background(), req)
-		if err != nil {
-			t.Fatalf("handleGetReportingExceptions returned error: %v", err)
-		}
-		if !result.IsError {
-			t.Error("Expected error result for missing lei parameter")
-		}
-	})
-}
-
-// TestGetHandler tests the getHandler dispatcher.
-func TestGetHandler(t *testing.T) {
-	registry := newTestRegistry("http://unused")
-
-	tests := []struct {
-		name     string
-		toolName string
-		wantNil  bool
-	}{
-		{"lei_lookup", "lei_lookup", false},
-		{"validate_lei", "validate_lei", false},
-		{"batch_lei_lookup", "batch_lei_lookup", false},
-		{"search_entity", "search_entity", false},
-		{"search_by_bic", "search_by_bic", false},
-		{"search_by_isin", "search_by_isin", false},
-		{"search_by_country", "search_by_country", false},
-		{"get_relationships", "get_relationships", false},
-		{"autocomplete", "autocomplete", false},
-		{"get_lei_issuer", "get_lei_issuer", false},
-		{"list_lei_issuers", "list_lei_issuers", false},
-		{"get_reporting_exceptions", "get_reporting_exceptions", false},
-		{"unknown_tool", "nonexistent_tool", false}, // Returns error handler
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			handler := registry.lookupHandler(tt.toolName)
-			if handler == nil && !tt.wantNil {
-				t.Errorf("lookupHandler(%q) returned nil", tt.toolName)
-			}
-		})
-	}
-}
-
-// TestUnknownToolHandler tests that unknown tools return an error.
-func TestUnknownToolHandler(t *testing.T) {
-	registry := newTestRegistry("http://unused")
-	handler := registry.lookupHandler("nonexistent_tool")
-	req := makeRequest(map[string]any{})
-
-	result, err := handler(context.Background(), req)
-	if err != nil {
-		t.Fatalf("Unknown tool handler returned error: %v", err)
-	}
-	if !result.IsError {
-		t.Error("Expected error result for unknown tool")
-	}
-	text := getResultText(result)
-	if !strings.Contains(text, "Unknown tool") {
-		t.Errorf("Expected 'Unknown tool' in error message, got: %s", text)
-	}
-}
-
 // TestParseArgumentsEmptyRequest tests parseArguments with empty request.
 func TestParseArgumentsEmptyRequest(t *testing.T) {
 	req := &mcp.CallToolRequest{
